[REFACTOR] authService에서 getUserId 메소드 분리 및 패키지 구조 변경 내역 반영
