+++ bankingService/src/main/kotlin/com/example/bankingservice/controller/AuthController.kt
@@ -3,9 +3,8 @@ package com.example.bankingservice.controller
 import com.example.bankingservice.controller.dto.ApiResponse
 import com.example.bankingservice.controller.dto.LoginRequestDto
 import com.example.bankingservice.controller.dto.LoginResponseDto
-import com.example.bankingservice.controller.dto.SignUpRequestDto
 import com.example.bankingservice.global.auth.JwtTokenProvider
-import com.example.bankingservice.service.AuthService
+import com.example.bankingservice.service.auth.AuthService
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.security.core.AuthenticationException
