+++ bankingService/src/main/kotlin/com/example/bankingservice/service/auth/AuthService.kt
@@ -1,8 +1,7 @@
-package com.example.bankingservice.service
+package com.example.bankingservice.service.auth
 
 import org.springframework.security.core.Authentication
 
 interface AuthService {
     fun authenticate(username: String, password: String): Authentication
-    fun getUserId(): Long
 }
