+++ bankingService/src/main/kotlin/com/example/bankingservice/service/user/impl/FriendListServiceImpl.kt
@@ -1,8 +1,9 @@
-package com.example.bankingservice.service
+package com.example.bankingservice.service.user.impl
 
 import com.example.bankingservice.domain.friendList.FriendList
 import com.example.bankingservice.domain.friendList.FriendListRepository
 import com.example.bankingservice.domain.user.User
+import com.example.bankingservice.service.user.FriendListService
 import org.springframework.stereotype.Service
 
 @Service
