+++ bankingService/src/test/kotlin/com/example/bankingservice/UserServiceTest.kt
@@ -2,17 +2,13 @@ package com.example.bankingservice
 
 import com.example.bankingservice.domain.user.User
 import com.example.bankingservice.domain.user.repository.UserRepository
-import com.example.bankingservice.service.UserService
+import com.example.bankingservice.service.user.UserService
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase
-import org.springframework.boot.test.autoconfigure.orm.jpa.AutoConfigureTestEntityManager
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.test.context.ActiveProfiles
 import org.springframework.test.context.junit.jupiter.SpringExtension
-import org.springframework.transaction.annotation.Transactional
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
 import kotlin.test.assertTrue
